fix: ensure all Person instances have valid ids
addresses #812
