+++ renku/core/management/datasets.py
@@ -35,10 +35,10 @@ import requests
 
 from renku.core import errors
 from renku.core.management.config import RENKU_HOME
-from renku.core.models.datasets import Creator, Dataset, DatasetFile, \
-    DatasetTag
+from renku.core.models.datasets import Dataset, DatasetFile, DatasetTag
 from renku.core.models.git import GitURL
 from renku.core.models.locals import with_reference
+from renku.core.models.provenance.agents import Person
 from renku.core.models.refs import LinkReference
 
 
@@ -391,7 +391,7 @@ class DatasetsApiMixin(object):
                 creators = []
                 # grab all the creators from the commit history
                 for commit in repo.iter_commits(paths=path):
-                    creator = Creator.from_commit(commit)
+                    creator = Person.from_commit(commit)
                     if creator not in creators:
                         creators.append(creator)
