+++ renku/core/management/repository.py
@@ -394,8 +394,8 @@ class RepositoryApiMixin(GitCore):
         self.repo.description = name or path.name
 
         # Check that an creator can be determined from Git.
-        from renku.core.models.datasets import Creator
-        Creator.from_git(self.repo)
+        from renku.core.models.provenance.agents import Person
+        Person.from_git(self.repo)
 
         # TODO read existing gitignore and create a unique set of rules
         import pkg_resources
