+++ tests/core/commands/test_serialization.py
@@ -18,8 +18,13 @@
 """Serialization tests for renku models."""
 
 import datetime
+from urllib.parse import quote, urljoin
 
 import yaml
+
+from renku.core.models.datasets import Dataset
+from renku.core.models.jsonld import NoDatesSafeLoader
+from renku.core.utils.doi import is_doi
 
 
 def test_dataset_serialization(client, dataset, data_file):
@@ -73,12 +78,6 @@ def test_dataset_deserialization(client, dataset):
 
 def test_doi_migration(doi_dataset):
     """Test migration of id with doi."""
-    import yaml
-    from urllib.parse import quote, urljoin
-    from renku.core.utils.doi import is_doi
-    from renku.core.models.datasets import Dataset
-    from renku.core.models.jsonld import NoDatesSafeLoader
-
     dataset = Dataset.from_jsonld(
         yaml.load(doi_dataset, Loader=NoDatesSafeLoader)
     )
@@ -88,3 +87,15 @@ def test_doi_migration(doi_dataset):
         'https://localhost', 'datasets/' + quote(dataset.identifier, safe='')
     ) == dataset._id
     assert dataset.same_as == urljoin('https://doi.org', dataset.identifier)
+
+
+def test_dataset_creator_email(doi_dataset):
+    """Check that creators without an email are assigned a blank node."""
+    # modify the dataset metadata to change the creator
+    dataset = Dataset.from_jsonld(
+        yaml.load(doi_dataset, Loader=NoDatesSafeLoader)
+    )
+
+    dataset.creator[0]._id = 'mailto:None'
+    dataset_broken = Dataset.from_jsonld(dataset.asjsonld())
+    assert 'mailto:None' not in dataset_broken.creator[0]._id
